introduce executionSummaryCollector
