+++ dbms/src/Flash/Coprocessor/DAGResponseWriter.h
@@ -18,6 +18,7 @@
 #include <DataTypes/IDataType.h>
 #include <Flash/Coprocessor/ChunkCodec.h>
 #include <Flash/Coprocessor/DAGQuerySource.h>
+#include <Flash/Coprocessor/ExecutionSummary.h>
 #include <common/logger_useful.h>
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wunused-parameter"
@@ -26,39 +27,6 @@
 
 namespace DB
 {
-/// do not need be thread safe since it is only used in single thread env
-struct ExecutionSummary
-{
-    UInt64 time_processed_ns;
-    UInt64 num_produced_rows;
-    UInt64 num_iterations;
-    UInt64 concurrency;
-    ExecutionSummary()
-        : time_processed_ns(0)
-        , num_produced_rows(0)
-        , num_iterations(0)
-        , concurrency(0)
-    {}
-
-    void merge(const ExecutionSummary & other, bool streaming_call)
-    {
-        if (streaming_call)
-        {
-            time_processed_ns = std::max(time_processed_ns, other.time_processed_ns);
-            num_produced_rows = std::max(num_produced_rows, other.num_produced_rows);
-            num_iterations = std::max(num_iterations, other.num_iterations);
-            concurrency = std::max(concurrency, other.concurrency);
-        }
-        else
-        {
-            time_processed_ns = std::max(time_processed_ns, other.time_processed_ns);
-            num_produced_rows += other.num_produced_rows;
-            num_iterations += other.num_iterations;
-            concurrency += other.concurrency;
-        }
-    }
-};
-
 class DAGResponseWriter
 {
 public:
