+++ dbms/src/TestUtils/MPPTaskTestUtils.cpp
@@ -11,10 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include <DataStreams/TiRemoteBlockInputStream.h>
 #include <TestUtils/MPPTaskTestUtils.h>
-
-#include <cstddef>
 
 namespace DB::tests
 {
@@ -85,19 +82,7 @@ ColumnsWithTypeAndName MPPTaskTestUtils::exeucteMPPTasks(QueryTasks & tasks, con
 {
     auto res = executeMPPQueryWithMultipleContext(properties, tasks, server_config_map);
     auto cols = readBlocks(res);
-    auto * exchange_receiver_stream = dynamic_cast<ExchangeReceiverInputStream *>(res[0].get());
-    assert(exchange_receiver_stream);
-    std::cout << "ywq test summary..." << std::endl;
-    std::cout << "summary size:" << exchange_receiver_stream->execution_summaries.size() << std::endl;
-    for (size_t i = 0; i < exchange_receiver_stream->execution_summaries.size(); ++i)
-    {
-        for (auto kv : exchange_receiver_stream->execution_summaries[i])
-        {
-            std::cout << kv.first << std::endl;
-            std::cout << kv.second.time_processed_ns << ", num produced rows: " << kv.second.num_produced_rows << ", num_iterations: " << kv.second.num_iterations << ", concurrency: " << kv.second.concurrency << std::endl;
-        }
-    }
-    
+    ExecutionSummaryCollector::instance().collect(res);
     return cols;
 }
